[2127] Cleaned up  refactored files package path
